Refactored code into feature packages to organize a bit more.
