+++ app/src/main/java/tigerhacks/android/tigerhacksapp/prizes/PrizeCardView.java
@@ -1,4 +1,4 @@
-package tigerhacks.android.tigerhacksapp;
+package tigerhacks.android.tigerhacksapp.prizes;
 
 import android.content.Context;
 import android.support.transition.TransitionManager;
