+++ app/src/main/java/tigerhacks/android/tigerhacksapp/schedule/ScheduleCardView.java
@@ -1,4 +1,4 @@
-package tigerhacks.android.tigerhacksapp;
+package tigerhacks.android.tigerhacksapp.schedule;
 
 import android.content.Context;
 import android.support.transition.TransitionManager;
