+++ app/src/main/java/tigerhacks/android/tigerhacksapp/service/network/TigerHacksService.java
@@ -1,8 +1,11 @@
-package tigerhacks.android.tigerhacksapp;
+package tigerhacks.android.tigerhacksapp.service.network;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Url;
+import tigerhacks.android.tigerhacksapp.prizes.PrizeList;
+import tigerhacks.android.tigerhacksapp.schedule.ScheduleItemList;
+import tigerhacks.android.tigerhacksapp.sponsors.SponsorList;
 
 /**
  * Created by Conno on 9/22/2018.
