+++ app/src/main/java/tigerhacks/android/tigerhacksapp/sponsors/Mentor.java
@@ -1,4 +1,4 @@
-package tigerhacks.android.tigerhacksapp;
+package tigerhacks.android.tigerhacksapp.sponsors;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
