+++ app/src/main/java/tigerhacks/android/tigerhacksapp/sponsors/SponsorList.java
@@ -1,8 +1,10 @@
-package tigerhacks.android.tigerhacksapp;
+package tigerhacks.android.tigerhacksapp.sponsors;
 
 import java.util.List;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
+
+import tigerhacks.android.tigerhacksapp.sponsors.Sponsor;
 
 public class SponsorList {
