+++ app/src/main/res/layout/fragment_map.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context="tigerhacks.android.tigerhacksapp.MapFragment">
+    tools:context="tigerhacks.android.tigerhacksapp.maps.MapFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
