+++ app/src/main/res/layout/fragment_sponsors.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="tigerhacks.android.tigerhacksapp.SponsorsFragment">
+    tools:context="tigerhacks.android.tigerhacksapp.sponsors.SponsorsFragment">
     <!-- TODO: Update blank fragment layout -->
 
     <ScrollView
