+++ app/src/main/res/layout/prize_card_layout.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<tigerhacks.android.tigerhacksapp.PrizeCardView xmlns:android="http://schemas.android.com/apk/res/android"
+<tigerhacks.android.tigerhacksapp.prizes.PrizeCardView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/prizeCard1"
@@ -86,4 +86,4 @@
 
         </LinearLayout>
     </LinearLayout>
-</tigerhacks.android.tigerhacksapp.PrizeCardView>
+</tigerhacks.android.tigerhacksapp.prizes.PrizeCardView>
